Move user to App\Models namespace
